Fix screenshots always capturing on save
This refactors the capturer function into an async function for easier
conditional reuse in `processData`.

A screenshot will only capture if the location is new, or if a
screenshot hasn't been captured for an existing location. Deleting a
screenshot will result in the capturer being called on the next location
save.